Added some more meat to the test plan.
Added more tests for sprint 1.
Still haven't written any text for the tests in sprint 1 as I'm not sure
if we are going to put the tests in the document for sprint 1 or as an
appendix at the very end of the report. In either case, the way I see it
there shouldn't be any real text in the sprint1_tests document outside
of the tables